update script + readme

+++ oms_compare.py
@@ -10,12 +10,14 @@ import glob
 
 BIG_DIC = {}
 THREADS = []
-TMP_WORKSPACE = "/tmp/langyu/"
+TMP_WORKSPACE = "/root/omscmd/"
 SUB_FILES_DIR = "sub_files/"
 SUB_FILE = "sub_file"
 RESULT_FILES_DIR = "result_files/"
-RESULT_SUB_DIFF_FILE = "result_diff_file"
-RESULT_SUB_SAME_FILE = "result_same_file"
+RESULT_SUB_DIFF_FILE = "result_sub_diff_file"
+RESULT_SUB_SAME_FILE = "result_sub_same_file"
+RESULT_DIFF_FILE = "result_diff_file"
+RESULT_SAME_FILE = "result_same_file"
 
 
 PARSER = OptionParser()
@@ -31,12 +33,19 @@ THREAD_NUM = options.thread_number
 SIZE_ENABLE = options.size_enable
 
 def clear_tmp_file():
-	s_files = glob.glob(TMP_WORKSPACE+SUB_FILES_DIR+"/*")
-	for f in s_files:
-		os.remove(f)
-	r_files = glob.glob(TMP_WORKSPACE+RESULT_FILES_DIR+"/*")
-	for f in r_files:
-		os.remove(f)
+	if os.path.exists(TMP_WORKSPACE+SUB_FILES_DIR) :
+		s_files = glob.glob(TMP_WORKSPACE+SUB_FILES_DIR+"/*")
+		for f in s_files:
+			os.remove(f)
+	else:
+		os.mkdir(TMP_WORKSPACE+SUB_FILES_DIR)
+
+	if os.path.exists(TMP_WORKSPACE+RESULT_FILES_DIR) :
+		r_files = glob.glob(TMP_WORKSPACE+RESULT_FILES_DIR+"/*")
+		for f in r_files:
+			os.remove(f)
+	else:	
+		os.mkdir(TMP_WORKSPACE+RESULT_FILES_DIR)
 
 # load the destfile into dic
 def generate_dst_big_dic(destfile):
@@ -88,17 +97,18 @@ def split_file(file_name, split_num):
 		with open(file_name) as file:
 			while i < int(split_num):
 				sub_file_name = TMP_WORKSPACE + SUB_FILES_DIR + SUB_FILE + str(i) + ".txt"
+				sub_f = None
 				try:
-					with open(sub_file_name, 'w') as sub_file:
+					with open(sub_file_name, 'w') as sub_f:
 						for num in range(sub_line):
 							line = file.readline()
 							if not line:
 								break
 							else:
-								sub_file.write(line)
+								sub_f.write(line)
 				finally:
-					if sub_file:
-						sub_file.close()				
+					if sub_f:
+						sub_f.close()				
 				i += 1
 				sub_file_list.append(sub_file_name)
 	finally:
@@ -127,10 +137,10 @@ def compare_object(file_name, dic, num):
 						else:
 							key = name+size
                 				#if(dic.has_key(key) and dic[key] < time):
-                				if(dic.has_key(key) and time >= dic[key]):
-							sub_same_file.write(name+'\n')
+                				if(dic.has_key(key) and time <= dic[key]):
+							sub_same_file.write(time + " " + name + " " + size +'\n')
 						else:
-							sub_diff_file.write(name+'\n')
+							sub_diff_file.write(time + " " + name + " " + size +'\n')
 		finally:
 			if tmp_file:
 				tmp_file.close()
@@ -143,6 +153,38 @@ def compare_object(file_name, dic, num):
 		if sub_same_file:
 			sub_same_file.close()
 
+def combine_result(number):
+	try:	
+		diff_file = open(TMP_WORKSPACE + RESULT_FILES_DIR + RESULT_DIFF_FILE +".txt",'w')
+		same_file = open(TMP_WORKSPACE + RESULT_FILES_DIR + RESULT_SAME_FILE +".txt",'w')
+		for i in range(int(number)):
+			tmp_sub_diff_file = None;
+			tmp_sub_same_file = None;
+			try:
+				with open(TMP_WORKSPACE + RESULT_FILES_DIR + RESULT_SUB_DIFF_FILE + str(i) +".txt","r") as tmp_sub_diff_file:
+					while 1:
+						lines = tmp_sub_diff_file.readlines()
+						if not lines:
+							break
+						for line in lines:
+							diff_file.write(line)
+				with open(TMP_WORKSPACE + RESULT_FILES_DIR + RESULT_SUB_SAME_FILE + str(i) + ".txt","r") as tmp_sub_same_file:
+					while 1:
+						lines = tmp_sub_same_file.readlines()
+						if not lines:
+							break
+						for line in lines:
+							same_file.write(line)
+			finally:
+				if tmp_sub_diff_file:
+					tmp_sub_diff_file.close()
+				if tmp_sub_same_file:
+					tmp_sub_same_file.close()
+	finally:
+		if diff_file:
+			diff_file.close()
+		if same_file:
+			same_file.close()
 
 if __name__ == '__main__':
 	#pdb.set_trace()
@@ -169,8 +211,9 @@ if __name__ == '__main__':
 	print str(datetime.datetime.now()) + "   finish starting the thread ..."
 	for k in THREADS:
 		k.join()
-	
 	print str(datetime.datetime.now()) + "   finish comparing object ..."
+	combine_result(len(THREADS))
+	print str(datetime.datetime.now()) + "   finish combining result sub files ..."
 	BIG_DIC.clear()
 	print str(datetime.datetime.now()) + "   finish clearing big dic ..."
 #	try:
